Fix class not being picked up on embedded images and so sizing wrong in some browsers
Class name on the embedded icons was wrong and so not being used. Some
browsers therefore ignored the size info ehich was not in the style but
the old way on the tag. Fixed the class naming. Also renamed all css
class name stubs from the old evidence hub naming to LITEMAP.