Delete connections when node removed from map
Add extra code to ry and delete connections for a node when the node is
removed from the map. Now doing this before the node remove is called as
well as inside the node remove function, as it was not always removing
connections and I am not sure why. So this is belt and braces really.
Should come back to this.